dockerizing app

+++ main.py
@@ -3,6 +3,7 @@ import os
 import shutil
 import zipfile
 from flask import Flask, request, Response, send_file, jsonify, make_response
+from src.utils.config import AppConfig
 from src.utils.download_from_source import Urls, downloadFromSource
 from src.utils.zipping import zip_to_result
 from datetime import datetime
@@ -18,14 +19,14 @@ def index():
 
         urls = Urls(**body)
 
+        timestamp = datetime.now().strftime("%Y%m%d%H%M%S")
+
+        final_file = os.path.join(os.getcwd(), f'result-{timestamp}{AppConfig.extension()}')
+
         downloadFromSource(urls=urls, i=len(urls.sources) - 1)
         
         response: Response = make_response()
         response.mimetype = 'application/zip'      
-        
-        timestamp = datetime.now().strftime("%Y%m%d%H%M%S")
-
-        final_file = os.path.join(os.getcwd(), f'result-{timestamp}.zip')
 
         response.headers["Content-Disposition"] = f"attachment; filename={final_file}"
 
@@ -38,11 +39,10 @@ def index():
         return response, 200
     
     except Exception as e:
-        return jsonify(e.args[0]), 500
+        return jsonify(e.args), 500
     
     finally:
-        os.remove(final_file)
-        shutil.rmtree(os.path.join(os.getcwd(), 'temp'))
-
-if __name__ == "__main__":
-    app.run(debug=True)
+        if os.path.isfile(final_file):
+            os.remove(final_file)
+        if os.path.isdir(AppConfig.temp_folder()):
+            shutil.rmtree(AppConfig.temp_folder())
